fix #6
